refactor / move all the league-raking dedicated components into a single folder
