+++ frontend/src/components/rankings/LeaderCard/LeaderCard.tsx
