+++ frontend/src/components/rankings/leagues/StandingsTable.tsx
@@ -1,4 +1,4 @@
-import type { Liga } from "../../types/league";
+import type { Liga } from "../../../types/league";
 type Props = { standings: Liga[] };
 export const StandingsTable = ({ standings }: Props) => (
   <p>StandingsTable — {standings.length} registres carregats</p>
